Don't always track `Not Now` tap as denied preflight permissions
If an onboarding user selects to share his contacts in the preflight dialog an then decides not to share them in the system alert we display the permission denied dialog. The `Not Now` button on the top right is owned by the `ShareContactsStepViewController` and would still trigger the tracking of the preflight analytics if tapped in this state.