Fix reactions not updating when a user first likes a message
The reactionsChangeInfo is `nil` in case the user has not reacted to a message before, this is caused by the way the observation internally works. The Message observer observers its reactions relation and adds a `ReactionObserver` to added reactions, this means that this `ReactionObserver` can not fire in case this reactions has not been added before. Because of this issue the UI should always check the `reactionsChanged` flag instead of the `reactionsChangeInfo` directly. We might address this issue on the data-model side by either creating a `ReactionChangeInfo` manually or making the `reactionsChangeInfo` private.