Add debug button to identify unread conversations
In some situation it’s not possible to find which conversations are causing the badge count to be non-zero, even if they all look read. This option in the debug menu will allow to identify which one is the offending conversation. Hopefully this will allow us to understand the pattern of this issue.