Add image thumbnail creation function taking max shorter side length instead of max length
When creating a thumbnail using CoreImage the max length is specified, however for non-square and especially panoramic images this means the longer side will have a maximum length of the specified value. In the collections view we display th eimages in a square format in image views using the aspect fill content mode, we thus want the shorter side of the image to be the one with the specified max length when creating a thumbnail. This PR adds a new method which takes a the shorter side length as parameter and calculates the max size based on the images aspect ratio.

As an example before when creating a max size 100pt thumbnail for a panomara with 4:1 aspect ratio the resulting height would be 25pt, now it will be 100pt.