refactor: Clean up analytics code (#5455)
* refactor: Remove code and analytics events related to registration/invites

On the spec we don't have any info to track events regarding registration. So it's safe to remove them for now. If we need them in the future, we can bring it back and translate them to the new way of doing things.

* refactor: remove calling the old tag event method from `.incoming` call state

In that call state we were calling the old tag method. In this case we remove it and we don't replace it with something else since we don't need to (according to the spec.

* refactor: Delete Analytics+StorableTrackingEvent

It's not being used anywhere. The tag method inside the Analytics extension isn't being called anywhere in the project. In the spec we don't have anything related to a StorableTrackingEvent. We can remove it for now.

* refactor: Delete CallQualitySurveyReview.swift & clean up

We delete the file and method tagCallQualityReview inside Analytics & Call Events. According to spec ain't needed and we don't need to collect such information. It's safe to assume that file and method(s) can be deleted

* refactor: Remove tagSettingsChanged from Analytics+Settings.swift & delete file

In addition to that we delete trackNewValue method in SettingsProperty.swift

* refactor: Delete Analytics+Registration.swift file & Registration folder

* refactor: Clean up Analytics+CallEvents file by deleting methods

* refactor: Delete Analytics+TeamInvites.swift

It is not needed anymore and it should have been removed a long time ago. Team Invite isn't being used anywhere in the project.

* refactor: Delete Backup Event enum and associated methods

According to the spec the product team don't want to track anything related to backup events. Since that's the case we remove everything around tracking such events.

In addition to that we remove tagDidRemoveService, track and tag(_ event: Event) method

* refactor: Delete setSuperProperty method from AnalyticsProvider

* refactor: Delete Analytics+CallEvents.swift

It's from the old way of doing things and it's not needed anymore.

* refactor: Delete Analytics+CallEvents.swift part 2

* refactor: Adding To-Do's to clean up things from Data model side

* refactor: Delete Analytics+Guests & Analytics+Services.swift

According to spec we don't need to track any events related to the files mentioned. Plus all the new code regarding the new way of doing analytics + Guests moved elsewhere.

* fix: linting issues

* chore: Add documentation for AppInfo

* chore: Add documentation for openingApp

* chore: Add documentation for contributed method

* chore: Add doc for failedToDecryptMessage method

* chore: Add doc for CallStateEvents methods

* fix: Typo in doc

Co-authored-by: John Nguyen <28632506+johnxnguyen@users.noreply.github.com>

* fix: SwiftLint trailing warnings

* fix: bring back tagEvent method since we're going to use it after all

Co-authored-by: John Nguyen <28632506+johnxnguyen@users.noreply.github.com>